ereader/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
watchdog==3.0.0
markdown==3.5.2
python-frontmatter==1.1.0
websockets==12.0
